feat(deleted-employee): enable bulk deletion of employees via checkbox selection [VG-268]

+++ views/products/product_list.php
@@ -50,10 +50,7 @@
                         <th onclick="sortTable(6)">
                             <div class="header-content">Quantity <i id="sortIconQuantity" class="fas fa-arrow-up small-icon"></i> <!-- Single arrow icon --></div>
                         </th>
-                        <?php
-                            if (isset($_SESSION['user_name']) && isset($_SESSION['user_role']) && $_SESSION['user_role'] === 'Admin'): ?>
-                                <th>Actions</th>
-                        <?php endif; ?>
+                        <th>Actions</th>
 
                     </tr>
 
@@ -86,9 +83,6 @@
                             <td><?php echo htmlspecialchars($product['barcode']); ?></td>
                             <td>$<?php echo number_format((float)$product['price'], 2, '.', ''); ?></td>
                             <td><?php echo htmlspecialchars($product['quantity']); ?></td>
-                            <?php
-                            if (isset($_SESSION['user_name']) && isset($_SESSION['user_role']) && $_SESSION['user_role'] === 'Admin'): ?>
-
                             <td class="text-center align-middle" style="width: 50px;">
                                 <div class="dropdown dropdown">
                                     <i class="bi bi-three-dots-vertical"
@@ -106,17 +100,19 @@
                                                 Edit
                                             </a>
                                         </li>
-                                        <li>
-                                            <a class="dropdown-item d-flex align-items-center gap-1 py-1 px-2 small text-danger"
-                                                href="/products/delete/<?= $product['id'] ?>" style="font-size: 0.8rem;">
-                                                <i class="bi bi-trash3" style="font-size: 0.8rem;"></i>
-                                                Delete
-                                            </a>
-                                        </li>
+                                        <?php
+                                        if (isset($_SESSION['user_name']) && isset($_SESSION['user_role']) && $_SESSION['user_role'] === 'Admin'): ?>
+                                            <li>
+                                                <a class="dropdown-item d-flex align-items-center gap-1 py-1 px-2 small text-danger"
+                                                    href="/products/delete/<?= $product['id'] ?>" style="font-size: 0.8rem;">
+                                                    <i class="bi bi-trash3" style="font-size: 0.8rem;"></i>
+                                                    Delete
+                                                </a>
+                                            </li>
+                                        <?php endif; ?>
                                     </ul>
                                 </div>
                             </td>
-                            <?php endif; ?>
                         </tr>
                     <?php endforeach; ?>
                 <?php else : ?>
